refactor: Extract parse_header to scraper/transcript_io.py, improve testability
- Move parse_header() and SEPARATOR to scraper/transcript_io.py so both
  build_site.py and generate_summaries.py import from a shared module
  instead of generate_summaries depending on build_site
- Make process_transcripts() accept optional transcripts_dir and
  summaries_dir params so tests can pass tmp_path directly instead of
  monkeypatching module globals
- Name the rate-limit constant _API_RATE_LIMIT_S = 1
- Update main.py to import SEPARATOR from scraper.transcript_io
- Update test imports accordingly
